# new-workflows/spec/dist/operations/createWorkflowRun/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "createWorkflowRun",
    "version": "1.0.0"
  },
  "paths": {
    "/api/workflows/{id}/runs": {
      "post": {
        "summary": "Start a run of a registered workflow",
        "operationId": "createWorkflowRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "A registered workflow's id, or a plan file's `path` (from GET /plans) to run a config-defined workflow.\n",
            "schema": {
              "type": "string",
              "example": "example/hello"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "input"
                ],
                "properties": {
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "mode": {
                    "$ref": "#/components/schemas/workflow-run-mode"
                  },
                  "agentConfig": {
                    "$ref": "#/components/schemas/workflow-run-agent-config"
                  },
                  "skipTodos": {
                    "type": "boolean"
                  },
                  "cwd": {
                    "type": "string",
                    "description": "Overrides the mount's default working directory (e.g. dev-site's repo checkout root) for this run.\n",
                    "example": "/home/user/my-project"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "run"
                  ],
                  "properties": {
                    "run": {
                      "$ref": "#/components/schemas/workflow-run"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No registered workflow with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "workflow-run-mode": {
        "type": "string",
        "description": "Execution mode for a run, set once at `createRun` and unchanged for its lifetime. Same modes as today's `@saflib/workflows`: `dry`/`checklist` never touch the filesystem or run commands; `script` runs mechanical steps (copy, command) but skips prompts, TODO checks, and validation commands (typecheck/test); `print` hands prompts to an external agent and halts between steps; `run` drives the agent itself.\n",
        "enum": [
          "dry",
          "script",
          "print",
          "run",
          "checklist"
        ],
        "example": "run"
      },
      "workflow-run-agent-config": {
        "type": "object",
        "description": "Which agent CLI drives a run's `run`-mode steps, and its resumable session. Set at `createRun` (or when the run is retried against a different agent) and persisted on the run row so a resumed process can `--resume` the same underlying session.\n",
        "additionalProperties": false,
        "required": [
          "cli"
        ],
        "properties": {
          "cli": {
            "type": "string",
            "description": "Agent CLI adapter to use. `mock-agent` is test-only (no subprocess).\n",
            "enum": [
              "cursor-agent",
              "claude-agent",
              "mock-agent"
            ],
            "example": "cursor-agent"
          },
          "sessionId": {
            "type": "string",
            "description": "Session id returned by the agent CLI's first turn, passed back on subsequent turns (e.g. `--resume`) so multi-turn steps share context.\n",
            "example": "a1b2c3d4"
          }
        },
        "example": {
          "cli": "cursor-agent",
          "sessionId": "a1b2c3d4"
        }
      },
      "workflow-run": {
        "type": "object",
        "description": "A single workflow run — the wire form of a `workflow_run` row (`new-workflows-db/schemas/workflow-run.ts`).\n",
        "additionalProperties": false,
        "required": [
          "id",
          "workflow_source",
          "workflow_ref",
          "input",
          "mode",
          "skip_todos",
          "status",
          "current_step_index",
          "cwd",
          "agent_config",
          "base_commit_hash",
          "completion_hash",
          "parent_run_id",
          "parent_step_index",
          "created_at",
          "updated_at",
          "is_advancing",
          "was_cancelled"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "0SG36FJe6m4W"
          },
          "workflow_source": {
            "type": "string",
            "enum": [
              "code",
              "config"
            ],
            "example": "code"
          },
          "workflow_ref": {
            "type": "string",
            "description": "A code workflow's `defineWorkflow` id, or a `workflow_config` row id.",
            "example": "example/hello"
          },
          "input": {
            "type": "object",
            "description": "The input this run was started with.",
            "additionalProperties": true
          },
          "mode": {
            "$ref": "#/components/schemas/workflow-run-mode"
          },
          "skip_todos": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/workflow-run-status"
          },
          "current_step_index": {
            "type": "integer",
            "example": 1
          },
          "cwd": {
            "type": "string",
            "example": "/home/user/my-project"
          },
          "agent_config": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/workflow-run-agent-config"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_commit_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "The repo's HEAD commit hash when this run was created, if `cwd` resolved inside a git repo (null otherwise). Bookends a run — pairs with a preview (hypothetical, from here forward) and, once the run is done, a reflection (actual, from here to the repo's current HEAD) — assuming nothing else committed to the repo meanwhile.\n",
            "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
          },
          "completion_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "The repo's HEAD commit hash the moment this run first reached `done` (captured once, then fixed — not a live lookup). Null until the run is done. Paired with `base_commit_hash` for the run's \"reflection\": what it actually changed, isolated from anything that happened before or after it.\n",
            "example": "b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3"
          },
          "parent_run_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set when this run was spawned by a `call-workflow` step in another run."
          },
          "parent_step_index": {
            "type": [
              "integer",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "is_advancing": {
            "type": "boolean",
            "description": "True while this run (or a nested call-workflow descendant advancing on its behalf) is actively being advanced right now, server-side — independent of `status`, which only reflects the last *completed* step. Lets a client that reloaded mid-step (losing its own \"request still pending\" state) tell \"actively running\" apart from \"idle\".\n"
          },
          "was_cancelled": {
            "type": "boolean",
            "description": "True when `status` is `failed` because a person stopped it (see the run's Stop button) rather than a genuine step error — there's no separate `WorkflowRunStatus` for this; cancelling just makes the in-flight step's own promise reject, landing as an ordinary `\"failed\"`. Always false for any other `status`.\n"
          }
        }
      },
      "workflow-run-status": {
        "type": "string",
        "description": "Lifecycle state of a run. `pending` before the first step runs; `running` after a step succeeds with more steps left; `awaiting_prompt`/ `awaiting_user` after a step hands control back to the caller (an agent prompt or a human action is needed before advancing again); `done`/ `failed` are terminal.\n",
        "enum": [
          "pending",
          "running",
          "awaiting_prompt",
          "awaiting_user",
          "done",
          "failed"
        ],
        "example": "running"
      },
      "error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A short, machine-readable error code, for when HTTP status codes are not sufficient."
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the error.",
            "example": "The requested resource could not be found."
          }
        }
      }
    }
  }
}